centralized groupHotfixVersions function
